requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==39.0.2
decorator==5.1.1
durations==0.3.3
idna==3.4
ipaddress==1.0.23
nanoid==2.0.0
numpy==1.23.2
option==2.1.0
py==1.11.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
retry==0.9.2
humanfriendly==10.0
virtualenv==20.25.0